backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
motor==3.3.2
pymongo==4.6.1
pydantic==2.6.1
pydantic-settings==2.2.0
python-dotenv==1.0.1
httpx==0.27.0
firebase-admin==6.4.0
razorpay==1.3.0
apscheduler==3.10.4
scikit-learn==1.4.0
xgboost==2.0.3
numpy==1.26.4
pandas==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiohttp==3.9.3
beautifulsoup4==4.12.3
feedparser==6.0.11
geopy==2.4.1
pytz==2024.1
slowapi==0.1.9
pytest==8.1.0
pytest-asyncio==0.23.5
